演習問題2-10答え合わせ終了

+++ 2-10Q/Triangle.java
@@ -2,10 +2,10 @@ class Triangle {
   public static void main(String[] args) {
     int height = Integer.parseInt(args[0]);
     int width = Integer.parseInt(args[1]);
-    int area;
+    double area;
 
     // 面積の計算
-    area = height * width / 2;
+    area = height * width / 2.0;
 
     // 高さ・幅・面積の表示
     System.out.println("高さ:" + height);
